+++ src/main/java/insane96mcp/iguanatweaksreborn/module/mining/MiningMisc.java
@@ -6,25 +6,19 @@ import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.LoadFeature;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
-import insane96mcp.insanelib.util.MCUtils;
-import net.minecraft.world.entity.ai.attributes.AttributeModifier;
-import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.DiggerItem;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.*;
+import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.properties.SlabType;
-import net.minecraftforge.common.ForgeMod;
-import net.minecraftforge.event.entity.EntityJoinLevelEvent;
 import net.minecraftforge.event.entity.player.PlayerEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.event.config.ModConfigEvent;
 import net.minecraftforge.registries.ForgeRegistries;
 
-import java.util.UUID;
-
 @Label(name = "Misc", description = "Various mining changes")
 @LoadFeature(module = Modules.Ids.MINING)
 public class MiningMisc extends Feature {
-
-	public static final UUID BLOCK_REACH_REDUCTION_UUID = UUID.fromString("bae34f6a-c58e-4622-b2ab-f1b89b73b781");
 
 	@Config
 	@Label(name = "Insta-Mine Silverfish", description = "Silverfish blocks will insta-mine like pre-1.17")
@@ -32,10 +26,6 @@ public class MiningMisc extends Feature {
 	@Config
 	@Label(name = "Insta-Mine Heads", description = "Heads will insta-break")
 	public static Boolean instaMineHeads = true;
-
-	@Config(min = -4, max = 0)
-	@Label(name = "Mining Range reduction", description = "Reduce the range at which players can interact with blocks")
-	public static Double miningRangeReduction = -1d;
 
 	@Config
 	@Label(name = "Faster slabs, stairs and walls", description = "Makes slabs, stairs and walls take less time to break")
@@ -72,13 +62,7 @@ public class MiningMisc extends Feature {
 			event.setNewSpeed(event.getOriginalSpeed() * 1.5f);
 	}
 
-	@SubscribeEvent
-	public void onPlayerJoinLevel(EntityJoinLevelEvent event) {
-		if (!this.isEnabled()
-				|| miningRangeReduction == 0d
-				|| !(event.getEntity() instanceof Player player))
-			return;
-
-		MCUtils.applyModifier(player, ForgeMod.BLOCK_REACH.get(), BLOCK_REACH_REDUCTION_UUID, "Block reach reduction", miningRangeReduction, AttributeModifier.Operation.ADDITION, false);
+	public static int destroyDelay(ItemStack stack, DiggerItem item, BlockState state) {
+		return isEnabled(MiningMisc.class) && efficiencyBasedDestroyDelay ? Math.max(5 - (int) (item.speed / 2f), 1) : 5;
 	}
 }